Move all types into `types.ts` file
Motivation: we should put all our types into a single file
